refactor(agent-chat): split composer editor hook seams
Reduce regression risk in the chat composer by isolating selection, autocomplete, and event orchestration logic behind focused modules and seam tests.
